requirements.txt
# --- Core Framework and Server ---
Flask==2.3.2
gunicorn==21.2.0

# --- AI/LLM Libraries ---
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
openai==1.64.0  # Updated from 1.30.1
tiktoken==0.9.0

# --- Environment and Config ---
python-dotenv==1.0.1  # Using 1.0.1 (from first list)
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
typing-inspect==0.9.0
typing_extensions==4.12.2

# --- Data/Database ---
pandas==2.2.3
numpy==1.26.4
SQLAlchemy==2.0.38

pyarrow==19.0.1

# --- Supabase/Storage ---
postgrest==0.19.3
realtime==2.4.1
storage3==0.11.3
supabase==2.13.0
supafunc==0.9.3

# --- HTTP/Networking ---
requests==2.32.3  # Using 2.32.3 (from first list)
requests-toolbelt==1.0.0
urllib3==2.3.0
multidict==6.1.0
yarl==1.18.3
httpx==0.27.0  # Added as a common dependency for 'openai' package

# --- Utility/Other ---
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.28.0
orjson==3.10.15
packaging==24.2
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
pydeck==0.9.1
Pygments==2.19.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rich==13.9.4
rpds-py==0.23.1
schedule==1.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.2
StrEnum==0.4.15
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
tzdata==2025.1
watchdog==6.0.0
websockets==14.2
zstandard==0.23.0